+++ Changelog.md
@@ -5,9 +5,12 @@
 - Class Sle\TYPO3\Extbase\Domain\Model\BaseMultiStepValueObject
 - Class Sle\TYPO3\Extbase\Domain\Model\BaseValueObject
 - Class Sle\TYPO3\Extbase\Domain\Utility\EntityUtility
+- Session Database methods: persist, fetch, delete
+- FeUser::getUserTSConfig()
 
 ### Changes
 - All FeUser methods changed into static methods
+- Remove Session::removeAll()
 
 ## v0.3.0
 ### Added